Game.js added; UserForm imported in App.js
